requirements.txt
appdirs==1.4.4
asgiref==3.7.0
astroid==2.3.3
colorama==0.4.3
distlib==0.3.1
Django==3.2.23
django-crispy-forms==1.8.1
django-filter==2.4.0
django-login-required-middleware==0.4
django-widget-tweaks==1.4.5
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
pytz>=2019.3
six==1.13.0
sqlparse==0.4.4
virtualenv==20.4.3
wrapt==1.11.2
django-simple-history==3.1.1
num2words==0.5.12
gunicorn==20.0.4
# psycopg2==2.9.10
validate_docbr==1.10.0
python-dotenv==0.13.0
xhtml2pdf==0.2.8
reportlab==4.0.9
xlsxwriter==3.2.0
whitenoise==5.0.1
geopy==2.4.1
geographiclib==2.0